# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.RegisterInput:
    properties:
      email:
        maxLength: 254
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 32
        type: string
    required:
    - email
    - password
    - username
    type: object
  gin.H:
    additionalProperties: {}
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@servasec.local
    name: API Support
  description: Security scanner management platform API
  title: Servasec API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with access_token and refresh_token cookies
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/gin.H'
      summary: Authenticate user
      tags:
      - Auth
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: User profile including features
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get current user
      tags:
      - Auth
    patch:
      consumes:
      - application/json
      parameters:
      - description: Fields to update (username, email)
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Profile updated
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update current user profile
      tags:
      - Auth
  /me/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: Current and new password
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Password updated
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Current password is incorrect
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update password
      tags:
      - Auth
...
  /refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed, new access_token cookie set
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Missing, invalid or blacklisted refresh token
          schema:
            $ref: '#/definitions/gin.H'
      summary: Refresh access token
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
        "400":
          description: Invalid input
          schema:
            type: object
        "409":
          description: Username or email already exists
          schema:
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"